Words to letters in LINQ queries
This is to appease Ryan, but also cuz like conventions. Went through and
changed everywhere it used a word (which also happened to be the real
variable's name) to a litter to use in the LINQ query.
This was done in both the tutorial content file and the code snippets
files.